# internal/index: batch calls to binary.Read to limit reflection overhead