gives about a 1.5-4X speed improvement in bam.ReadIndex, depending
on the index.