@brentp 2016-12-24
BTW, it's showing that I have permission to merge this PR before travis has even finished.