@brentp 2016-12-24
here's the benchcmp, these are pretty stable:
```
benchmark                old ns/op     new ns/op     delta
BenchmarkReadIndex-4     7282756       3993347       -45.17%
```

and for another: 
```
benchmark                old ns/op     new ns/op     delta
BenchmarkReadIndex-4     1864271       332785        -82.15%
```