@brentp 2016-12-24
After the change to bench:
```
benchmark                old ns/op     new ns/op     delta
BenchmarkReadIndex-4     1504545       860009        -42.84%
```

```
benchmark                old ns/op     new ns/op     delta
BenchmarkReadIndex-4     343927        177406        -48.42%
```