@kortschak 2016-12-26
Please repeat the benchmarks and use benchstat instead of benchcmp. Also, I'm wondering how much this actually impacts on a normal use. The index is read once and is generally fairly short. Do we actually need to do this?