@brentp 2016-12-26
Here is the benchmark with benchstats:

```
name         old time/op  new time/op  delta
ReadIndex-4   345µs ± 1%   178µs ± 3%  -48.53%  (p=0.008 n=5+5)
```

I have changed to to allocate the buffer inside of readChunks instead of passing it in as there is no speed difference between those 2 modes.

I am building an SV viewer for [bigly](https://github.com/brentp/bigly) and this helps it to be more responsive. It halves the time to read the index.