@kortschak 2016-12-26
>I am building an SV viewer for bigly and this helps it to be more responsive. It halves the time to read the index.

Yeah, I can understand that, I'm just wondering how much that is of the entire cost.