@brentp 2016-12-26
> Yeah, I can understand that, I'm just wondering how much that is of the entire cost.

Up to as must as one-half