@kortschak 2016-12-26
Really? So you're only reading index files then?