@brentp 2016-12-26
I think I have this all cleaned up. 

> Really? So you're only reading index files then?

I'm reading the index, and then parsing the bam file and running bigly pileup on a region. If that region is small enough then reading the index can be half of the total request time. This is true for a WGS bam index where the index itself can be quite large.
And, to clarify, I mean that the time to read the index was up to half of the total request time.