@brentp 2016-12-26
>> And, to clarify, I mean that the time to read the index was up to half of the total request time.

> OK, I'm asking about the total time spent to get from the input data to the desired result, which should include the BAM reading time and your analysis.

I haven't benchmarked that rigorously, just know that, to read the index of an example was on the order of a second and the entire request was on the order of 2 seconds